# FIREFLY-1796: Optimize Redis usage when querying jobs by user