--- a/src/firefly/java/edu/caltech/ipac/firefly/core/background/JobManager.java
+++ b/src/firefly/java/edu/caltech/ipac/firefly/core/background/JobManager.java
@@ -19,22 +19,21 @@ import edu.caltech.ipac.firefly.server.events.ServerEventManager;
 import edu.caltech.ipac.firefly.server.events.WebsocketConnector;
 import edu.caltech.ipac.firefly.server.util.Logger;
 import edu.caltech.ipac.util.AppProperties;
-import edu.caltech.ipac.util.cache.Cache;
 import edu.caltech.ipac.util.cache.CacheKey;
 import edu.caltech.ipac.util.cache.CacheManager;
 import edu.caltech.ipac.util.cache.StringKey;
 import org.apache.commons.lang.text.StrBuilder;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
+import redis.clients.jedis.params.ScanParams;
 
 import javax.annotation.Nonnull;
 import java.io.Serializable;
 import java.time.Duration;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
+import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Objects;
@@ -51,7 +50,6 @@ import static edu.caltech.ipac.firefly.data.ServerParams.EMAIL;
 import static edu.caltech.ipac.util.StringUtils.isEmpty;
 import static edu.caltech.ipac.firefly.core.background.Job.Type.PACKAGE;
 import static edu.caltech.ipac.firefly.core.background.JobInfo.Phase.*;
-import static java.util.Optional.ofNullable;
 
 
 /**
@@ -64,19 +62,22 @@ public class JobManager {
 
     public static final String BG_INFO = "background.info";
     public static final String ALL_JOB_CACHE_KEY = "ALL_JOB_INFOS"; // cache key for all job infos
-    public static final long CLEANUP_INTVL_MINS = AppProperties.getIntProperty("job.cleanup.interval", 60);     // run cleanup once every 60 minutes
+    public static final long CLEANUP_INTVL_MINS = AppProperties.getIntProperty("job.cleanup.interval", 12*60);          // run cleanup once every 12 hours
+    public static final int JOB_LIST_DEFAULT_LIMIT = AppProperties.getIntProperty("job.list.default.limit", 100_000);    // the default limit for job list queries
     private static final int KEEP_ALIVE_INTERVAL = AppProperties.getIntProperty("job.keepalive.interval", 60);  // default keepalive interval in seconds
     private static final int WAIT_COMPLETE = AppProperties.getIntProperty("job.wait.complete", 1);              // wait for complete after submit in seconds
     private static final int MAX_PACKAGERS = AppProperties.getIntProperty("job.max.packagers", 10);             // maximum number of simultaneous packaging threads
     private static final int JOB_EXPIRY_HOURS = AppProperties.getIntProperty("job.expiry.hours", 24*14);        // Time in hours to keep a job after it has ended.  Default to 14 days.
+    public static final int JOB_SCAN_BATCH_SIZE = AppProperties.getIntProperty("job.scan.batch_size", 10_000);   // batch size for scanning job keys in Redis.  Default to 10,000.  Larger value return more keys per call but use more CPU and memory per iteration.  this is a good size for larger redis store.
 
     private static final Logger.LoggerImpl LOG = Logger.getLogger();
     private static final ExecutorService packagers = Executors.newFixedThreadPool(MAX_PACKAGERS);
     private static final ExecutorService searches = Executors.newCachedThreadPool();
     private static final HashMap<String, JobEntry> runningJobs = new HashMap<>();
-    private static final Cache<JobInfo> allJobInfos = new DistribMapCache<>(ALL_JOB_CACHE_KEY, 0, new JobInfoSerializer()); // the all job hash should never expire
+    private static final DistribMapCache<JobInfo> allJobInfos = new DistribMapCache<>(ALL_JOB_CACHE_KEY, 0, new JobInfoSerializer()); // the all job hash should never expire
     private static final String COMPLETED_HANDLER = AppProperties.getProperty("job.completed.handler");
-
+    private static final CacheKey JOB_CACHE_VERSION_KEY = new StringKey("job.all.cache.version");
+    private static final String JOB_CACHE_VERSION = "1.0";
 
     static {
         if (!isEmpty(COMPLETED_HANDLER)) {
@@ -92,16 +93,26 @@ public class JobManager {
         Messenger.subscribe(JobEvent.TOPIC, new JobEventHandler());
         Executors.newSingleThreadScheduledExecutor().scheduleAtFixedRate(
                     JobManager::checkJobs, KEEP_ALIVE_INTERVAL, KEEP_ALIVE_INTERVAL, TimeUnit.SECONDS);   // check every 30 seconds
+
+
+        String jobCacheVersion = (String) CacheManager.getDistributed().get(JOB_CACHE_VERSION_KEY);
+        if (isEmpty(jobCacheVersion) || !jobCacheVersion.equals(JOB_CACHE_VERSION)) {
+            LOG.info("Migrating job cache keys to new format");
+            int count = migrateRedisKeys();
+            LOG.info("Migrated " + count + " job cache keys to new format");
+            CacheManager.getDistributed().put(JOB_CACHE_VERSION_KEY, JOB_CACHE_VERSION); // set the version
+        }
     }
 
     /**
      * @return a list of JobInfo belonging to the current request owner
      */
     public static List<JobInfo> list() {
+        List<JobInfo> userJobs = JobManager.getUserJobs();      // Ensure getUserJobs() is called only once, since it may be expensive. List is updated and returned at the end.
         UWS_HISTORY_SVCS.forEach(svc -> {
-            Try.it(() -> importJobHistories(svc)).getOrElse(LOG::error);
+            Try.it(() -> importJobHistories(svc, userJobs)).getOrElse(LOG::error);
         });
-        return getAllUserJobs();
+        return userJobs;
     }
 
     public static JobInfo submit(Job job) {
@@ -196,16 +207,27 @@ public class JobManager {
 //  Getter/Setter
 //====================================================================
 
+    public static JobInfo getJobInfo(String jobId) {
+        String userKey = ServerContext.getRequestOwner().getUserKey();
+        return getJobInfo(jobId, userKey);
+    }
+
     /**
-     * Retrieves the stored JobInfo with the given jobId.
+     * Retrieves the stored JobInfo with the given jobId and userKey.
      * Returned value is read-only.  Use updateJobInfo to update the JobInfo.
      * @param jobId the ID of the job
      * @return the JobInfo with the jobId, or null not found
      */
-    public static JobInfo getJobInfo(String jobId) {
+    public static JobInfo getJobInfo(String jobId, String userKey) {
         if (isEmpty(jobId)) return null;
-        return allJobInfos.get(cacheKey(jobId));
+        return allJobInfos.get(cacheKey(jobId, userKey));
     }
+
+    public static JobInfo getJobInfo(CacheKey ckey) {
+        if (isEmpty(ckey)) return null;
+        return allJobInfos.get(ckey);
+    }
+
 
     /**
      *  see {@link #updateJobInfo(String, boolean, Consumer)}
@@ -291,30 +313,40 @@ public class JobManager {
     public static String getStatistics(boolean details) {
         StrBuilder sb = new StrBuilder();
 
-
         sb.append        ("          |   Job Count Active Count  Error Count\n");
         sb.append        ("          |------------ ------------ ------------\n");
+        List<JobInfo> allJobs = getAllJobs();
 
         Arrays.stream(Job.Type.values()).forEach(type -> {
-            long total = getAllJobs().stream().filter(v -> v.getMeta().getType() == type).count();
-            long active = getAllJobs().stream().filter(v -> v.getPhase() == EXECUTING && v.getMeta().getType() == type).count();
-            long error = getAllJobs().stream().filter(v -> v.getPhase() == ERROR && v.getMeta().getType() == type).count();
+            long total = allJobs.stream().filter(v -> v.getMeta().getType() == type).count();
+            long active = allJobs.stream().filter(v -> v.getPhase() == EXECUTING && v.getMeta().getType() == type).count();
+            long error = allJobs.stream().filter(v -> v.getPhase() == ERROR && v.getMeta().getType() == type).count();
 
             sb.append(String.format("%9s |%,12d %,12d %,12d\n", type, total, active, error));
         });
 
         if (details) {
-            sb.append("\n");
-            sb.append("JOB ID               LOCAL JOB ID         PHASE       startTime              elapsedTime(s) progress monitored userKy                               \n");
+            allJobs.sort(
+                    Comparator.comparing(
+                            JobInfo::getCreationTime,
+                            Comparator.nullsFirst(Comparator.naturalOrder())
+                    ).reversed()
+            );
+
+
+            List<JobInfo> subList = allJobs.subList(0, 100);
+            sb.append("\n (ONLY THE LATEST 100 JOBS ARE SHOWN BELOW) \n");
+            sb.append("JOB ID               LOCAL JOB ID         PHASE       creationTime           elapsedTime(s) progress monitored userKy                               \n");
             sb.append("-------------------- -------------------- ----------- ---------------------- -------------- -------- --------- -------------------------------------\n");
-            getAllJobs().forEach((ji) -> {
-                Instant endT = ji.getEndTime() != null ? ji.getEndTime() : Instant.now();
+            subList.forEach((ji) -> {
+                Instant endT = ji.getEndTime();
+                Instant startT = ji.getStartTime();
                 sb.append(String.format("%-20s %-20s %-11s %-22s %,14d %8d %9s %-37s\n",
                         ji.getJobId(),
                         ji.getMeta().getJobId(),
                         ji.getPhase(),
-                        ji.getStartTime().truncatedTo(ChronoUnit.SECONDS),
-                        Duration.between(ji.getStartTime(), endT).toSeconds(),
+                        ji.getCreationTime() == null ? "" : ji.getCreationTime().truncatedTo(ChronoUnit.SECONDS),
+                        startT == null || endT == null ? -1 : Duration.between(startT, endT).toSeconds(),
                         ji.getMeta().getProgress(),
                         ji.getMeta().isMonitored(),
                         ji.getMeta().getUserKey()));
@@ -328,32 +360,31 @@ public class JobManager {
                 .map(je -> getJobInfo(je.job.getJobId()))               // convert JobEntry to JobInfo
                 .filter(Objects::nonNull).toList();                     // return only non-null JobInfo
     }
+    /**
+     * Get all job keys in the datastore.
+     * @return a list of all job keys in the datastore
+     */
+    static List<String> getAllJobKeys() {
+        return allJobInfos.getKeys().stream().map(CacheKey::getUniqueString).toList();
+    }
 
+    /**
+     * Get jobs history.  This may be expensive when there are huge number of jobs, so use with caution.
+     * @return a list of all JobInfo in the datastore
+     */
     static List<JobInfo> getAllJobs() {
-        List<? extends CacheKey> keys = allJobInfos.getKeys();
-        if (keys == null || keys.isEmpty()) return Collections.emptyList();
-        ArrayList<JobInfo> rval = new ArrayList<JobInfo>(keys.size());
-        for (CacheKey key : keys) {
-            JobInfo jobInfo = ifNotNull(() -> allJobInfos.get(key)).get();  // ignore error
-            if (jobInfo != null) {
-                rval.add(jobInfo);
-            }
-        }
-        return rval;
+        ScanParams scanParams = new ScanParams().match("*").count(JOB_SCAN_BATCH_SIZE); // adjust count as needed
+        return allJobInfos.getValuesFor(scanParams);
     }
 
     /**
      * Get all jobs that belong to the current user
      * @return a list of JobInfo that belong to the current user
      */
-    static List<JobInfo> getAllUserJobs() {
+    static List<JobInfo> getUserJobs() {
         String userKey = ServerContext.getRequestOwner().getUserKey();
-        return ofNullable(getAllJobs())
-                .orElse(Collections.emptyList())
-                .stream()
-                .filter(info -> userKey.equals(info.getMeta().getUserKey()) && info.getMeta().isMonitored())  // only return monitored jobs belonging to the current user
-                .toList();
-
+        ScanParams scanParams = new ScanParams().match("*:%s".formatted(userKey)).count(JOB_SCAN_BATCH_SIZE); // adjust count as needed
+        return allJobInfos.getValuesFor(scanParams);
     }
 
 //====================================================================
@@ -386,20 +417,27 @@ public class JobManager {
     }
 
     private static CacheKey cacheKey(JobInfo jobInfo) {
-        return cacheKey(jobInfo.getMeta().getJobId());
+        return cacheKey(jobInfo.getMeta().getJobId(), jobInfo.getMeta().getUserKey());
     }
 
-    private static CacheKey cacheKey(String jobId) {
-        return isEmpty(jobId) ? null : new StringKey(jobId);
+    /**
+     * Redis key is compose of jobId and userKey.  This is so that we can filter the keys for a specific user.
+     * jobId should still be unique.  If jobId is empty, it will return null.
+     * @param jobId
+     * @param userKey
+     * @return a CacheKey for the jobId and userKey, or null if jobId is empty
+     */
+    private static CacheKey cacheKey(String jobId, String userKey) {
+        return isEmpty(jobId) ? null : new StringKey(jobId + ":" + userKey);
     }
 
     private static void checkJobs() {
 
-        // ping clients with active(EXECUTING) job
-        getAllJobs().stream().filter(fi -> fi != null && fi.getPhase() == EXECUTING)     // all running jobs
-                .map(fi -> fi.getMeta().getUserKey() + "::" + fi.getMeta().getEventConnId()).distinct()       // get distinct list of userKey and connId
+        // ping clients with active(EXECUTING) job to keep the connection alive
+        getRunningJobs().stream().filter(fi -> fi != null && fi.getPhase() == EXECUTING)     // all running jobs
+                .map(fi -> fi.getMeta().getUserKey() + ":" + fi.getMeta().getEventConnId()).distinct()       // get distinct list of userKey and connId
                 .forEach(client -> {                                                    // ensure it gets pinged only once, regardless of the number of jobs it has.
-                    String[] ownerConnId = client.split("::");
+                    String[] ownerConnId = client.split(":");
                     WebsocketConnector.pingClient(ownerConnId[0], ownerConnId[1]);      // this will only ping client with the given owner/connId
                 });
 
@@ -499,8 +537,9 @@ public class JobManager {
     }
 
     public static void cleanup() {
-        allJobInfos.getKeys().forEach( k -> {
-            JobInfo job = allJobInfos.get(k);
+        List<JobInfo> jobs = getAllJobs();
+        jobs.forEach(job -> {
+            CacheKey k = cacheKey(job);
             if (!job.getMeta().isMonitored() && job.getEndTime().plus(1, ChronoUnit.HOURS).isBefore(Instant.now())) {
                 LOG.info("Removing non-monitored job: " + k);
                 allJobInfos.remove(k);      // remove non-monitored job after 1 hour
@@ -515,7 +554,7 @@ public class JobManager {
      * This serializer is used to serialize JobInfo objects for storage in the cache.
      * Instead of using the default Java serialization, it uses a JSON string.
      */
-    private static class JobInfoSerializer implements DistribMapCache.Serializer {
+    private static class JobInfoSerializer implements DistribMapCache.Serializer<JobInfo> {
 
         public String serialize(Object obj) {
             if (obj instanceof JobInfo jobInfo) {
@@ -524,11 +563,36 @@ public class JobManager {
             return null;
         }
 
-        public Object deserialize(String str) throws Exception{
+        public JobInfo deserialize(String str) throws Exception{
             if (str != null) {
                 return toJobInfo((JSONObject) new JSONParser().parse(str));
             }
             return null;
         }
+    }
+
+
+//====================================================================
+//  Utilities: adhoc use cases
+//====================================================================
+
+    /**
+     * Migrate all Redis keys that do not have userKey in the key to the new format.
+     * @return the number of keys migrated
+     */
+    public static int migrateRedisKeys() {
+        // For keys without userKey, we need to migrate them to the new format where userKey is appended to the jobId.
+        int count = 0;
+        for(CacheKey key : allJobInfos.getKeys()) {
+            if (key instanceof StringKey sk && !sk.getUniqueString().contains(":")) { // has userKey in the key
+                JobInfo jobInfo = allJobInfos.get(key);
+                if (jobInfo != null) {
+                    allJobInfos.put(cacheKey(jobInfo), jobInfo);
+                    allJobInfos.remove(key); // remove old key
+                    count++;
+                }
+            }
+        }
+        return count;
     }
 }