Ticket: https://jira.ipac.caltech.edu/browse/FIREFLY-1796

### Optimizations Implemented

* **Improved `getUserJobs`**: now uses a compound key to filter Redis keys by user before fetching `JobInfo`, significantly reducing lookup overhead.
* **Added `job_cache` versioning**: helps identify when a migration is required.
* **Migration support**: added logic to convert existing keys to compound-key format.
* **Minimized `getAllJobs` usage**:

  * This expensive call is now only used in `admin/status`.
  * Removed from other parts of the code, especially pinging logic.
* **Optimized `listUserJob`**:

  * Calls `getUserJobs()` once, updates results as needed, and returns the final list.
  * Returns full `JobInfo`, avoiding additional server calls for job details.
  * Supports `LAST` parameter to limit the number of jobs returned (default: 500 via config).
  * Also implemented `PHASE` and `AFTER` filters for future use.
* **Improved `importJobHistory`**:

  * Now receives `getUserJobs()` result as input.
  * Only fetches full `JobInfo` from UWS services for **active jobs** or if the job is not found in Firefly's history.

---

### Testing Notes

* **Job Cache loaded with 101,104 jobs** (mocked).

* **Performance Benchmarks**:

  ```
  getAllJobs: 9,341ms with 101,104 jobs
  getUserJobs: 10,668ms with 99,994 jobs
  getUserJobs:    221ms with 1,000 jobs
  getUserJobs:    105ms with   100 jobs
  getUserJobs:     90ms with    10 jobs
  ```

* `getAllJobs()` (9.3s) is now only used in `admin/status`.

* `getUserJobs()` takes only **221ms** for 1,000 jobs.

---

### Manual Test

* Start Firefly.
* Inject the 1,000 job userKey into the browser cookie.
* Open **Job Monitor** → only 500 jobs returned (default limit), loaded quickly with **no additional network call**.

---

### Redis Health

```
redis.rdb: 39 MB
RAM usage: 106 MB
```